HvD: The screening on gammaval was too harsh. For benzene in the "DZVP (DFT orbital)" basis the results for becke97 with and without screening on gammaval were:
                xbecke97            hfexch+cbecke97
with screening  -230.468473394454   -232.412470395985
no screening    -230.468503072263   -232.412446368920

These differences are maybe not physically significant but they are large
enough to very likely cause concern amongst users. Hence I have tightened 
the screening up (in analogy to screeing on gammaval in other functionals)
and now the results are essentially indistinguishable from the no screening
results.

This change affects all varieties of the Becke'97 and HCTH functionals.



git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@27068 85dd5168-0e29-45a6-bdf9-2122c44e486f