add F90 allocatable in addition to MA for buffers
MA is still the default - F90 guarded by preprocessor.

F90 (de)allocate are required to enable KNL MCDRAM because attributes on the arrays is how the compiler knows to put them there.  the alternative is to make API change to MA, which is wildly disruptive on the rest of the code, and we do not want that.

the argument for MA is because buffers need to be registered with the NIC for RDMA.  in the case here, we allocate only once per function call and thus GA can register the buffer on the first use and it will be negligible overhead.  and many networks will not even require registration for the local buffers.  it is only on the remote side of RDMA where it matters.

the preprocessor guarding is ugly and will eventually be fixed via the method noted in the comment in the code.



git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@27224 85dd5168-0e29-45a6-bdf9-2122c44e486f